motor_statusをint型にして微調節を可能に

+++ AKROS_bridge_converter/include/AKROS_bridge_converter/motor_status.h
@@ -5,14 +5,14 @@
 typedef struct motor_status_{
     uint8_t CAN_ID;
     bool servo_mode;    // サーボ状態
-    float position;
-    float position_ref;
-    float velocity;
-    float velocity_ref;
-    float effort;
-    float effort_ref;
-    float Kp;
-    float Kd;
+    uint16_t position;
+    uint16_t position_ref;
+    uint16_t velocity;
+    uint16_t velocity_ref;
+    uint16_t effort;
+    uint16_t effort_ref;
+    uint16_t Kp;
+    uint16_t Kd;
 }motor_status;
 
 #endif
